requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
absl-py==1.0.0; python_version >= '3.6'
addfips==0.3.1
aiohttp==3.8.1
aiosignal==1.2.0; python_version >= '3.6'
appnope==0.1.2; platform_system == 'Darwin'
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2; python_version >= '3.6'
attrs==21.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
backcall==0.2.0
cachetools==5.0.0; python_version ~= '3.7'
certifi==2021.10.8
charset-normalizer==2.0.12; python_version >= '3.5'
click==8.0.4; python_version >= '3.6'
cycler==0.11.0; python_version >= '3.6'
debugpy==1.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
decorator==5.1.1; python_version >= '3.5'
entrypoints==0.4; python_version >= '3.6'
executing==0.8.3
flask==2.0.3
flatbuffers==2.0
fonttools==4.31.1; python_version >= '3.7'
frozenlist==1.3.0; python_version >= '3.7'
gast==0.5.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-auth-oauthlib==0.4.6; python_version >= '3.6'
google-auth==2.6.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-pasta==0.2.0
grpcio==1.44.0; python_version >= '3.6'
h5py==3.6.0; python_version >= '3.7'
idna==3.3; python_version >= '3'
importlib-metadata==4.11.3; python_version < '3.10'
importlib-resources==5.4.0; python_version >= '3.6'
ipykernel==6.9.2
ipython==8.1.1; python_version >= '3.8'
itsdangerous==2.1.1; python_version >= '3.7'
jedi==0.18.1; python_version >= '3.6'
jinja2==3.0.3; python_version >= '3.6'
jsonschema==4.4.0; python_version >= '3.7'
jupyter-client==7.1.2; python_full_version >= '3.6.1'
jupyter-core==4.9.2; python_version >= '3.6'
keras-preprocessing==1.1.2
keras==2.8.0
kiwisolver==1.4.0; python_version >= '3.7'
libclang==13.0.0
markdown==3.3.6; python_version >= '3.6'
markupsafe==2.1.1; python_version >= '3.7'
matplotlib-inline==0.1.3; python_version >= '3.5'
matplotlib==3.5.1
more-itertools==8.12.0
multidict==6.0.2; python_version >= '3.7'
nbformat==5.2.0
nest-asyncio==1.5.4; python_version >= '3.5'
numpy==1.22.3
oauthlib==3.2.0; python_version >= '3.6'
opt-einsum==3.3.0; python_version >= '3.5'
packaging==21.3; python_version >= '3.6'
pandas==1.4.1
parso==0.8.3; python_version >= '3.6'
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
pillow==9.0.1; python_version >= '3.7'
plotly==5.6.0
prompt-toolkit==3.0.28; python_full_version >= '3.6.2'
protobuf==3.19.4; python_version >= '3.5'
psutil==5.9.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1-modules==0.2.8
pyasn1==0.4.8
pygments==2.11.2; python_version >= '3.5'
pyparsing==3.0.7; python_version >= '3.6'
pyrsistent==0.18.1; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.1
pyzmq==22.3.0; python_version >= '3.6'
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.27.1
rsa==4.8; python_version >= '3.6'
setuptools==60.10.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
stack-data==0.2.0
tenacity==8.0.1; python_version >= '3.6'
tensorboard-data-server==0.6.1; python_version >= '3.6'
tensorboard-plugin-wit==1.8.1
tensorboard==2.8.0; python_version >= '3.6'
tensorflow-io-gcs-filesystem==0.24.0; python_version < '3.11' and python_version >= '3.7'
tensorflow==2.8.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tornado==6.1; python_version >= '3.5'
traitlets==5.1.1; python_version >= '3.7'
typing-extensions==4.1.1; python_version >= '3.6'
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
wcwidth==0.2.5
werkzeug==2.0.3; python_version >= '3.6'
wheel==0.37.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
wrapt==1.14.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
xlrd==2.0.1
yarl==1.7.2; python_version >= '3.6'
zipp==3.7.0; python_version < '3.10'
